backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "groq>=1.2.0",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "uvicorn[standard]>=0.46.0",
]
